solved day9

+++ day2/src/main.rs
@@ -1,6 +1,6 @@
 extern crate intcode;
 fn main() {  
-  let memory: Vec<i32> = std::fs::read_to_string("./input.txt")
+  let memory: Vec<i64> = std::fs::read_to_string("./input.txt")
     .unwrap()
     .split(",")
     .map(|x| x.parse().unwrap())
@@ -12,18 +12,18 @@ fn main() {
   println!("part 2: {}", part2_result);
 }
 
-fn set_verb_noun(memory: &mut Vec<i32>, verb: i32, noun: i32) -> &mut Vec<i32> {
+fn set_verb_noun(memory: &mut Vec<i64>, verb: i64, noun: i64) -> &mut Vec<i64> {
   memory[1] = verb;
   memory[2] = noun;
   return memory;
 }
 
-fn part1(memory: &mut Vec<i32>) -> i32 {
+fn part1(memory: &mut Vec<i64>) -> i64 {
   let (result, _, _) = intcode::run_program(set_verb_noun(memory, 12, 2).clone(), || 0, |_| {});
   return result;
 }
 
-fn part2(memory: &mut Vec<i32>) -> i32{
+fn part2(memory: &mut Vec<i64>) -> i64{
 let search_for = 19690720;
   let mut noun = 0;
   let mut verb = 0;
